requirements.txt
pandas==2.3.1
python-dotenv==1.1.1
